# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_Tencent_WeKnora_internal_errors.AppError:
    properties:
      code:
        $ref: '#/definitions/github_com_Tencent_WeKnora_internal_errors.ErrorCode'
      details: {}
      message:
        type: string
    type: object
  github_com_Tencent_WeKnora_internal_errors.ErrorCode:
    enum:
    - 1000
    - 1001
    - 1002
    - 1003
    - 1004
    - 1005
    - 1006
    - 1007
    - 1008
    - 1009
    - 1010
    - 2000
    - 2001
    - 2002
    - 2003
    - 2004
    - 2100
    - 2101
    - 2102
    - 2103
    type: integer
    x-enum-varnames:
    - ErrBadRequest
    - ErrUnauthorized
    - ErrForbidden
    - ErrNotFound
    - ErrMethodNotAllowed
    - ErrConflict
    - ErrTooManyRequests
    - ErrInternalServer
    - ErrServiceUnavailable
    - ErrTimeout
    - ErrValidation
    - ErrTenantNotFound
    - ErrTenantAlreadyExists
    - ErrTenantInactive
    - ErrTenantNameRequired
    - ErrTenantInvalidStatus
    - ErrAgentMissingThinkingModel
    - ErrAgentMissingAllowedTools
    - ErrAgentInvalidMaxIterations
    - ErrAgentInvalidTemperature
  github_com_Tencent_WeKnora_internal_types.AnswerConfidenceEvidenceItem:
    properties:
      chunk_id:
        type: string
      current_feedback:
        type: string
      id:
        type: string
      knowledge_base_id:
        type: string
      knowledge_id:
        type: string
      match_type:
        type: string
      position:
        type: integer
      rerank_score:
        type: number
      retrieval_score:
        type: number
      source_channel:
        type: string
      source_type:
        type: string
      title:
        type: string
    type: object
  github_com_Tencent_WeKnora_internal_types.AnswerConfidenceResponse:
    properties:
      confidence_label:
        type: string
      confidence_score:
        type: number
      evidence_status:
        type: string
      evidence_strength_label:
        type: string
      evidence_strength_score:
        type: number
      evidences:
        items:
          $ref: '#/definitions/github_com_Tencent_WeKnora_internal_types.AnswerConfidenceEvidenceItem'
        type: array
      message_id:
        type: string
      reference_count:
        type: integer
      source_count:
        type: integer
      source_health_label:
        type: string
      source_health_score:
        type: number
      source_type_counts:
        additionalProperties:
          type: integer
        type: object
    type: object
...
  internal_handler.SourceFeedbackRequest:
    properties:
      comment:
        type: string
      evidence_id:
        type: string
      feedback:
        type: string
    required:
    - evidence_id
    - feedback
    type: object
info:
  contact:
    name: WeKnora Github
    url: https://github.com/Tencent/WeKnora
  description: WeKnora 知识库管理系统 API 文档
  termsOfService: http://swagger.io/terms/
  title: WeKnora API
  version: "1.0"
paths:
  /chat/answer/{message_id}/confidence:
    get:
      consumes:
      - application/json
      description: Returns evidence strength, source health, and cited evidence details for an assistant answer
      parameters:
      - description: Assistant message ID
        in: path
        name: message_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Confidence details
          schema:
            properties:
              data:
                $ref: '#/definitions/github_com_Tencent_WeKnora_internal_types.AnswerConfidenceResponse'
              success:
                type: boolean
            type: object
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/github_com_Tencent_WeKnora_internal_errors.AppError'
        "404":
          description: Message not found
          schema:
            $ref: '#/definitions/github_com_Tencent_WeKnora_internal_errors.AppError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_Tencent_WeKnora_internal_errors.AppError'
      security:
      - Bearer: []
      - ApiKeyAuth: []
      summary: Get answer confidence details
      tags:
      - chat
  /chat/answer/{message_id}/feedback:
    post:
      consumes:
      - application/json
      description: Submits up, down, or expired feedback for a cited source in an assistant answer
      parameters:
      - description: Assistant message ID
        in: path
        name: message_id
        required: true
        type: string
      - description: Source feedback payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handler.SourceFeedbackRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Feedback accepted
          schema:
            properties:
              success:
                type: boolean
            type: object
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/github_com_Tencent_WeKnora_internal_errors.AppError'
        "404":
          description: Message or evidence not found
          schema:
            $ref: '#/definitions/github_com_Tencent_WeKnora_internal_errors.AppError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_Tencent_WeKnora_internal_errors.AppError'
      security:
      - Bearer: []
      - ApiKeyAuth: []
      summary: Submit source feedback
      tags:
      - chat
...
securityDefinitions:
  ApiKeyAuth:
    description: 租户身份认证：输入 sk- 开头的 API Key
    in: header
    name: X-API-Key
    type: apiKey
  Bearer:
    description: 用户登录认证：输入 Bearer {token} 格式的 JWT 令牌
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"